Pull up following revision(s) (requested by christos in ticket #1398): 	sys/netsmb/subr_mchain.c: revision 1.19 via patch 	sys/netsmb/smb_subr.c: revision 1.35 	sys/netsmb/smb_subr.h: revision 1.19 	sys/netsmb/mchain.h: revision 1.9 via patch 	sys/netsmb/smb_dev.h: revision 1.7 Convert sizes/lengths to unsigned (size_t) or uint32_t (for binary compatibility).

+++ sys/netsmb/smb_subr.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: smb_subr.h,v 1.16 2006/08/17 17:11:29 christos Exp $	*/
+/*	$NetBSD: smb_subr.h,v 1.16.6.1 2010/07/22 20:36:22 riz Exp $	*/
 
 /*
  * Copyright (c) 2000-2001, Boris Popov
@@ -109,18 +109,18 @@ void smb_makescred(struct smb_cred *scred, struct lwp *l,
     kauth_cred_t cred);
 int  smb_proc_intr(struct lwp *);
 char *smb_strdup(const char *s);
-char *smb_strdupin(char *s, int maxlen);
-void *smb_memdupin(void *umem, int len);
+char *smb_strdupin(char *s, size_t maxlen);
+void *smb_memdupin(void *umem, size_t len);
 void smb_strtouni(u_int16_t *dst, const char *src);
 void smb_strfree(char *s);
 void smb_memfree(void *s);
-void *smb_zmalloc(unsigned long size, struct malloc_type *type, int flags);
+void *smb_zmalloc(size_t size, struct malloc_type *type, int flags);
 
 int  smb_encrypt(const u_char *apwd, u_char *C8, u_char *RN);
 int  smb_ntencrypt(const u_char *apwd, u_char *C8, u_char *RN);
 int  smb_maperror(int eclass, int eno);
 int  smb_put_dmem(struct mbchain *mbp, struct smb_vc *vcp,
-	const char *src, int len, int caseopt);
+	const char *src, size_t len, int caseopt);
 int  smb_put_dstring(struct mbchain *mbp, struct smb_vc *vcp,
 	const char *src, int caseopt);
 int  smb_put_string(struct smb_rq *rqp, const char *src);
